refactor: renamed view to presentation and file changes
