+++ lib/app/app.dart
@@ -1,8 +1,1 @@
-// Copyright (c) 2022, Very Good Ventures
-// https://verygood.ventures
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-export 'view/app.dart';
+export 'presentation/app.dart';
