+++ lib/app/presentation/app.dart
@@ -4,10 +4,10 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:style_guide_infra/style_guide_infra.dart';
 
 import '../../cubit/selected_style_guide/selected_style_guide_cubit.dart';
-import '../../home/view/home_page.dart';
+import '../../home/presentation/home_page.dart';
 import '../../l10n/l10n.dart';
 import '../../routes.dart';
-import '../cubit/all_style_guide_cubit.dart';
+import 'cubit/all_style_guide_cubit.dart';
 
 class App extends StatelessWidget {
   const App({super.key});
