+++ lib/app/presentation/cubit/all_style_guide_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
 
 import '../../../cubit/selected_style_guide/selected_style_guide_cubit.dart';
-import '../model/style_guide_model.dart';
+import '../../model/style_guide_model.dart';
 
 part 'all_style_guide_state.dart';
