+++ lib/home/home.dart
@@ -1,1 +1,1 @@
-export 'view/home_page.dart';
+export 'presentation/home_page.dart';
